small changes in SessionAnalysis

+++ ZPI_Task_two/src/data_analysis/SessionAnalysis.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 
 public class SessionAnalysis {
-    static double getNumberOfIncreasingSessions(ArrayList<Double> data){
+    static int getNumberOfIncreasingSessions(ArrayList<Double> data){
         int res = 0;
         for(int i = 1; i < data.size(); i++){
             if(data.get(i)>data.get(i-1))
@@ -13,7 +13,7 @@ public class SessionAnalysis {
         return res;
     }
 
-    static double getNumberOfDecreasingSessions(ArrayList<Double> data){
+    static int getNumberOfDecreasingSessions(ArrayList<Double> data){
         int res = 0;
         for(int i = 1; i < data.size(); i++){
             if(data.get(i)<data.get(i-1))
@@ -22,7 +22,7 @@ public class SessionAnalysis {
         return res;
     }
 
-    static double getNumberOfSessionsWithoutChange(ArrayList<Double> data){
+    static int getNumberOfSessionsWithoutChange(ArrayList<Double> data){
         int res = 0;
         for(int i = 1; i < data.size(); i++){
             if(data.get(i).equals(data.get(i - 1)))
